feat: Spawns Addon editor

+++ src/app/services/query.service.ts
@@ -145,16 +145,16 @@ export class QueryService {
 
   // Tracks difference between two groups of rows (with TWO keys) and generate DELETE/INSERT query
   getDiffDeleteInsertTwoKeysQuery<T extends TableRow>(
-    tableName: string,   // the name of the table (example: 'creature_loot_template')
-    primaryKey1: string, // first  primary key (example: 'Entry')
-    primaryKey2: string, // second primary key (example: 'Item')
-    currentRows: T[],    // object of the original rows
-    newRows: T[],        // array of the new rows
+    tableName: string,        // the name of the table (example: 'creature_loot_template')
+    primaryKey1: string|null, // first  primary key (example: 'Entry')
+    primaryKey2: string,      // second primary key (example: 'Item')
+    currentRows: T[],         // object of the original rows
+    newRows: T[],             // array of the new rows
   ): string {
 
     if (!newRows || !currentRows) { return ''; }
 
-    if (newRows.length === 0) {
+    if (primaryKey1 && newRows.length === 0) {
       // all rows have been deleted
       return `DELETE FROM \`${tableName}\` WHERE \`${primaryKey1}\` = ${currentRows[0][primaryKey1]};\n`;
     }
@@ -172,8 +172,11 @@ export class QueryService {
     const deleteQuery: squel.Delete = squel.delete(squelConfig).from(tableName);
     const insertQuery: squel.Insert = squel.insert(squelConfig).into(tableName);
 
-    deleteQuery.where('`' + primaryKey1 + '` = ' + newRows[0][primaryKey1]);
+    if (primaryKey1) {
+      deleteQuery.where('`' + primaryKey1 + '` = ' + newRows[0][primaryKey1]);
+    }
     deleteQuery.where('`' + primaryKey2 + '` IN ?', involvedRows);
+
     insertQuery.setFieldsRows(addedOrEditedRows);
 
     return this.getFinalDiffDeleteInsertQuery(addedOrEditedRows, deleteQuery, insertQuery);
@@ -211,15 +214,22 @@ export class QueryService {
   getFullDeleteInsertQuery<T extends TableRow>(
     tableName: string,          // the name of the table (example: 'creature_loot_template')
     rows: T[],                  // array of the new rows
-    primaryKey: string,         // first primary key (example: 'Entry'), it will be used to generate the DELETE statement
-    primaryKey2: string = null, // (optional) the second primary key, it will be used to generate the DELETE statement
+    primaryKey: string = null,  // first primary key (example: 'Entry'), it will be used to generate the DELETE statement for ALL rows
+    primaryKey2: string = null, // the second primary key, it will be used to generate the DELETE statement for SPECIFIC rows
   ) {
     if (!rows || rows.length === 0) { return ''; }
 
-    let deleteCondition: string = '`' + primaryKey + '` = ' + rows[0][primaryKey];
+    let deleteCondition: string = '';
+
+    if (primaryKey) {
+      deleteCondition += '`' + primaryKey + '` = ' + rows[0][primaryKey];
+    }
+    if (primaryKey && primaryKey2) {
+      deleteCondition += ` AND `;
+    }
     if (primaryKey2) {
       const ids = rows.map(row => row[primaryKey2]);
-      deleteCondition += ' AND `' + primaryKey2 + '` IN (' + ids.join(', ') + ')';
+      deleteCondition += '`' + primaryKey2 + '` IN (' + ids.join(', ') + ')';
     }
 
     const deleteQuery: squel.Delete = squel.delete(squelConfig).from(tableName).where(deleteCondition);
